NO-SNOW: Fix DDL rows_affected; share query-type taxonomy in sf_core
DDL statements previously returned a misleading rows_affected=1 because
calculate_rows_affected blindly passed through Snowflake's data.total
(populated with 1 as a generic success marker) for any non-DML statement.

Lift the statementTypeId taxonomy (QueryType + ResultKind) out of the
odbc crate into sf_core::query_types::statement_type as the single source
of truth, and key rows_affected off the shared classifier:

- UpdateCount (DML): sum affected-row columns from the rowset (unchanged).
- Cursor (SELECT/SHOW/...): the result-set size in data.total.
- NoResult (DDL/TCL/unknown): None, instead of the misleading data.total.

odbc/src/api/query_type.rs is now a thin re-export of the sf_core types,
so the native result path and the ODBC SQLRowCount path classify
statements identically and can no longer drift.

- query_types.rs -> query_types/ module (row_type.rs unchanged, history
  preserved via git mv; new statement_type.rs; mod.rs re-exports both).
- Remove duplicated DML constants / is_dml_statement from result_set.rs.
- Tests: add DDL-is-None and unknown-is-None unit tests, fix the SELECT
  test to use a real SELECT statementTypeId, and update the ALTER SESSION
  integration assertions from Some(1) to None.
