+++ sf_core/src/apis/database_driver_v1/result_set.rs
@@ -6,6 +6,9 @@ use super::global_state::{DatabaseDriverV1, WrapperPresets};
 use super::query::build_reader_from_rowset_data;
 use crate::chunks::{ChunkDownloadData, ChunkFormatKind, PrefetchConfig};
 use crate::handle_manager::Handle;
+use crate::query_types::statement_type::{
+    DML_AFFECTED_ROWS_COLUMN_PREFIXES, DML_AFFECTED_ROWS_COLUMNS, QueryType, ResultKind,
+};
 use crate::rest::snowflake::query_response::{Data, RowType, RowsetData, Stats};
 use crate::rest::snowflake::snowflake_get_query_result;
 use arrow::array::RecordBatchReader;
@@ -140,73 +143,47 @@ pub(super) struct ResultSet {
     pub reader_ctx: ReaderContext,
 }
 
-// --- DML detection constants ---
-
-const DML_AFFECTED_ROWS_COLUMNS: &[&str] = &[
-    "number of rows updated",
-    "number of multi-joined rows updated",
-    "number of rows deleted",
-];
-const DML_AFFECTED_ROWS_COLUMN_PREFIXES: &[&str] = &["number of rows inserted"];
-
-const STATEMENT_TYPE_ID_DML: i64 = 0x3000;
-const STATEMENT_TYPE_ID_INSERT: i64 = 0x3100;
-const STATEMENT_TYPE_ID_UPDATE: i64 = 0x3200;
-const STATEMENT_TYPE_ID_DELETE: i64 = 0x3300;
-const STATEMENT_TYPE_ID_MERGE: i64 = 0x3400;
-const STATEMENT_TYPE_ID_MULTI_TABLE_INSERT: i64 = 0x3500;
-const STATEMENT_TYPE_ID_GET_FILES: i64 = 0x7101;
-const STATEMENT_TYPE_ID_PUT_FILES: i64 = 0x7102;
-
 // --- Response parsing helpers ---
 
-fn is_dml_statement(statement_type_id: Option<i64>) -> bool {
-    statement_type_id.is_some_and(|type_id| {
-        matches!(
-            type_id,
-            STATEMENT_TYPE_ID_DML
-                | STATEMENT_TYPE_ID_INSERT
-                | STATEMENT_TYPE_ID_UPDATE
-                | STATEMENT_TYPE_ID_DELETE
-                | STATEMENT_TYPE_ID_MERGE
-                | STATEMENT_TYPE_ID_MULTI_TABLE_INSERT
-        )
-    })
+/// Calculate rows affected from a query response, keyed off the statement's
+/// [`ResultKind`] (the shared `query_types::statement_type` classifier).
+///
+/// - `UpdateCount` (DML): sum the affected-row columns from the rowset.
+/// - `Cursor` (SELECT / SHOW / ...): the result-set size in `data.total`.
+/// - `NoResult` (DDL / TCL / unknown): `None`. Snowflake returns `total: 1` as a
+///   generic success marker for these; surfacing it as `rows_affected = 1` is
+///   misleading, so we report "not applicable" instead.
+pub(super) fn calculate_rows_affected(data: &Data) -> Option<i64> {
+    match QueryType::from_raw(data.statement_type_id).result_kind() {
+        ResultKind::UpdateCount => Some(sum_dml_affected_rows(data)),
+        ResultKind::Cursor => data.total,
+        ResultKind::NoResult => None,
+    }
 }
 
-/// Calculate rows affected based on statement type.
-///
-/// Returns `Some(count)` when rows affected is known, `None` when it is not
-/// (when the statement type is unknown).
-///
-/// - For DML: Parse rowset columns to sum affected rows
-/// - For SELECT and other queries: Use total field
-/// - For unknown: Return None
-pub(super) fn calculate_rows_affected(data: &Data) -> Option<i64> {
-    if is_dml_statement(data.statement_type_id) {
-        if let (Some(rowset), Some(row_types)) = (&data.rowset, &data.row_type)
-            && !rowset.is_empty()
-            && !rowset[0].is_empty()
-        {
-            let mut affected_rows = 0i64;
-            for (idx, col) in row_types.iter().enumerate() {
-                let col_name = col.name.to_lowercase();
-                if (DML_AFFECTED_ROWS_COLUMNS.contains(&col_name.as_str())
-                    || DML_AFFECTED_ROWS_COLUMN_PREFIXES
-                        .iter()
-                        .any(|p| col_name.starts_with(p)))
-                    && let Some(Some(value)) = rowset[0].get(idx)
-                    && let Ok(count) = value.parse::<i64>()
-                {
-                    affected_rows += count;
-                }
-            }
-            return Some(affected_rows);
-        }
-        return Some(0);
+/// Sum the integer cells of the DML affected-row columns in the first rowset row.
+fn sum_dml_affected_rows(data: &Data) -> i64 {
+    let (Some(rowset), Some(row_types)) = (&data.rowset, &data.row_type) else {
+        return 0;
+    };
+    if rowset.is_empty() || rowset[0].is_empty() {
+        return 0;
     }
 
-    data.total
+    let mut affected_rows = 0i64;
+    for (idx, col) in row_types.iter().enumerate() {
+        let col_name = col.name.to_lowercase();
+        if (DML_AFFECTED_ROWS_COLUMNS.contains(&col_name.as_str())
+            || DML_AFFECTED_ROWS_COLUMN_PREFIXES
+                .iter()
+                .any(|p| col_name.starts_with(p)))
+            && let Some(Some(value)) = rowset[0].get(idx)
+            && let Ok(count) = value.parse::<i64>()
+        {
+            affected_rows += count;
+        }
+    }
+    affected_rows
 }
 
 pub(super) fn response_to_descriptor(
@@ -227,8 +204,8 @@ pub(super) fn response_to_descriptor(
         .unwrap_or_default();
 
     let statement_type_id = data.statement_type_id.or(match data.command.as_deref() {
-        Some("UPLOAD") => Some(STATEMENT_TYPE_ID_PUT_FILES),
-        Some("DOWNLOAD") => Some(STATEMENT_TYPE_ID_GET_FILES),
+        Some("UPLOAD") => Some(QueryType::PUT_FILES.raw()),
+        Some("DOWNLOAD") => Some(QueryType::GET_FILES.raw()),
         _ => None,
     });
