+++ sf_core/src/query_types/mod.rs
@@ -1,0 +1,5 @@
+mod row_type;
+pub mod statement_type;
+
+pub use row_type::{GeoRepresentation, RowType, VectorElementType};
+pub use statement_type::{QueryType, ResultKind};
